client user deserialization
